currentcolor support for color-mix
color-mix values are generally resolved into single blink::Colors at
computed value time. "currentcolor" cannot do this as its value is
unknown until after style is computed.

In the event that a color-mix contains "currentcolor" this CL adds
a private UnresolvedColorMix attribute to StyleColor that will get
resolved at used value time when currentcolor is available.

If neither color is a "currentcolor" we continue to resolve color-mix
values at computed value time.

It is still unresolved what we'll do when the value of both percentages
for color-mix are zero as it should get rejected at parse time but
there are no WPT tests that cover this situation yet.

Bug: 1333988
Change-Id: I954f2371bc267e7912b62dbaccc363e5bdf901da