Set requestStart/responseStart to connectEnd for prefetch.
For speculation rule based prefetch, the resource-fetching request
occurs before the navigation, resulting negative requestStart and responseStart. Set both negative values to connectEnd to respect the
`<=` timeline events ordering.

Notice that this might not be the spec'ed behavior of these two values.
Follow-up action is required to re-examine the behaviors when the spec
is updated.

Relevant doc:
https://docs.google.com/document/d/1wMelfFY34jnyCDA6AAy0nm3l16vE0ruTo3y3rnEaso4 [internal access].

Bug: 1382255
Change-Id: I6e25b9304074217c6618a12c2aafbbe32a7cf51e