Change LCP startTime to DomHRTime.
Change time conversion for LCP.

This is to make the conversion of LCP start time from TimeTicks to
double in the same way as that of FCP start time does so that the
precision of the two is the same. Currently they are not. The time conversion for LCP happens in WindowPerformance::OnLargestContentfulPaintUpdated(). It converts MonotonicTime to TimeDelta in milliseconds while FCP time is from MonotonicTime to DOMHighResTimeStamp.

This is found during investigation for crbug.com/1345948 where we want to align LCP start time with that of FCP.

See screenshots in the associated bug.

Bug: 1382187
Change-Id: Ibcf88dd61965669e7bf76b905fa2279f80119429