Add testing of declarative invocation of combined popovers
There weren't any tests of declarative invocation (e.g. via the
`popovertoggletarget` attribute) on "combined" elements such as
`<dialog popover>`. This adds them.

Along the way, this exposed a missing check for declarative invokers.
Because InvokePopover() calls showPopover(ASSERT_NO_EXCEPTION), it
needs to pre-check all of the same conditions that showPopover checks.
For this reason, the pre-check conditions were refactored into a
helper, which is now called for declarative invocations.

Bug: 1307772
Change-Id: I26bc797030fce36fe78845c09dfb218520aad6bc