[mathml] Force writing-mode to horizontal-tb in UA sheet
Currently Blink forces writing-mode to horizontal-tb in the style
adjuster, but this does not work well with logical properties [1]. This
CL move the logic to the UA sheet, as it is currently done in the MathML
Core specification [2]. WPT tests are added to check that writing-mode
cannot be overridden on MathML elements and that an author-specified
writing-mode does not affect logical properties. Indentation errors for
the universal rules are also fixed.

[1] https://phabricator.services.mozilla.com/D48279#5004118
[2] https://w3c.github.io/mathml-core/#user-agent-stylesheet

Bug: 6606
Change-Id: I5a109a9161a572aab7d94edc561bd9445b1f0182